Move query param logic into SearchContext.
